Strategy for keys and JSDoc
