+++ src/tests/rate-limit.test.ts
@@ -7,7 +7,7 @@ import type { ServiceInstances } from "../types";
 import { HPKEService } from "../services/security/hpke.service";
 import { TrustedService } from "../services/security/trusted.service";
 import { InMemoryOTPService } from "../services/security/otp/otp.service";
-import { KeyService } from "../services/security/key.service";
+import { KeyService } from "../services/user/key.service";
 import type { EmailService } from "../services/communication/email.service";
 import type { MetricsService } from "../services/metrics.service";
 import { TeeKeyService } from "../services/security/tee-key.service";
@@ -16,126 +16,126 @@ import { FPEService } from "../services/security/fpe.service";
 
 // Mock email service that doesn't actually send emails
 class MockEmailService implements EmailService {
-	async sendOTPEmail(
-		otp: string,
-		recipient: string,
-		projectName: string,
-		expiryMinutes?: string,
-		projectLogo?: string,
-	): Promise<void> {
-		return Promise.resolve();
-	}
+  async sendOTPEmail(
+    otp: string,
+    recipient: string,
+    projectName: string,
+    expiryMinutes?: string,
+    projectLogo?: string
+  ): Promise<void> {
+    return Promise.resolve();
+  }
 }
 
 // Mock metrics service that doesn't actually send metrics
 class MockMetricsService implements MetricsService {
-	gauge(name: string, value: number, tags?: string[]): void {}
-	distribution(name: string, value: number, tags?: string[]): void {}
-	histogram(name: string, value: number, tags?: string[]): void {}
-	increment(name: string, value = 1, tags?: string[]): void {}
+  gauge(name: string, value: number, tags?: string[]): void {}
+  distribution(name: string, value: number, tags?: string[]): void {}
+  histogram(name: string, value: number, tags?: string[]): void {}
+  increment(name: string, value = 1, tags?: string[]): void {}
 
-	async flush(): Promise<void> {
-		return Promise.resolve();
-	}
+  async flush(): Promise<void> {
+    return Promise.resolve();
+  }
 }
 
 async function initializeServicesWithMockEmail(
-	identityKey: CryptoKeyPair,
+  identityKey: CryptoKeyPair
 ): Promise<ServiceInstances> {
-	const keyPairProvider = TeeKeyService.getInstance();
-	const encryptionService = new HPKEService(keyPairProvider);
-	const fpeService = new FPEService(keyPairProvider);
-	const symmetricEncryptionService = new AesGcmService(keyPairProvider);
+  const keyPairProvider = TeeKeyService.getInstance();
+  const encryptionService = new HPKEService(keyPairProvider);
+  const fpeService = new FPEService(keyPairProvider);
+  const symmetricEncryptionService = new AesGcmService(keyPairProvider);
 
-	const otpService = InMemoryOTPService.getInstance();
-	const emailService = new MockEmailService();
-	const keyService = new KeyService(identityKey);
-	const trustedService = new TrustedService(
-		otpService,
-		emailService,
-		keyService,
-		encryptionService,
-		fpeService,
-	);
-	const metricsService = new MockMetricsService();
+  const otpService = InMemoryOTPService.getInstance();
+  const emailService = new MockEmailService();
+  const keyService = new KeyService(identityKey);
+  const trustedService = new TrustedService(
+    otpService,
+    emailService,
+    keyService,
+    encryptionService,
+    fpeService
+  );
+  const metricsService = new MockMetricsService();
 
-	return {
-		teeKeyService: keyPairProvider,
-		trustedService,
-		symmetricEncryptionService,
-		metricsService,
-		encryptionService,
-	};
+  return {
+    teeKeyService: keyPairProvider,
+    trustedService,
+    symmetricEncryptionService,
+    metricsService,
+    encryptionService,
+  };
 }
 
 describe("Rate Limiting Integration", () => {
-	it("should allow 3 onboarding requests but return 429 on the 4th", async () => {
-		const identityKey = await devIdentityKey();
-		const services = await initializeServicesWithMockEmail(identityKey);
-		const app = await createApp(services);
+  it("should allow 3 onboarding requests but return 429 on the 4th", async () => {
+    const identityKey = await devIdentityKey();
+    const services = await initializeServicesWithMockEmail(identityKey);
+    const app = await createApp(services);
 
-		// Test data - using same signerId and authId for all requests to trigger rate limiting
-		const testData = {
-			deviceId: "test-device-123",
-			signerId: "test-signer-123",
-			projectName: "Test Project",
-			authId: "email:test@example.com",
-			encryptionContext: {
-				publicKey:
-					"BMxzex+oSRC3yiAPhYmIHrZ1iuVLmDO05O6x6+Q8oCO4Vi1Gd7NnOSyCpv6MwA6bZo4pR9ky4tEsRdKsVrR6eas=",
-			},
-		};
+    // Test data - using same signerId and authId for all requests to trigger rate limiting
+    const testData = {
+      deviceId: "test-device-123",
+      signerId: "test-signer-123",
+      projectName: "Test Project",
+      authId: "email:test@example.com",
+      encryptionContext: {
+        publicKey:
+          "BMxzex+oSRC3yiAPhYmIHrZ1iuVLmDO05O6x6+Q8oCO4Vi1Gd7NnOSyCpv6MwA6bZo4pR9ky4tEsRdKsVrR6eas=",
+      },
+    };
 
-		// First 3 requests should succeed (200 status)
-		for (let i = 1; i <= 3; i++) {
-			const req = new Request("http://localhost/v1/signers/start-onboarding", {
-				method: "POST",
-				headers: {
-					"Content-Type": "application/json",
-					authorization: env.ACCESS_SECRET,
-				},
-				body: JSON.stringify({
-					...testData,
-					deviceId: `${testData.deviceId}-${i}`, // Different device IDs
-				}),
-			});
+    // First 3 requests should succeed (200 status)
+    for (let i = 1; i <= 3; i++) {
+      const req = new Request("http://localhost/v1/signers/start-onboarding", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+          authorization: env.ACCESS_SECRET,
+        },
+        body: JSON.stringify({
+          ...testData,
+          deviceId: `${testData.deviceId}-${i}`, // Different device IDs
+        }),
+      });
 
-			const res = await app.fetch(req);
-			expect(res.status).toBe(200);
+      const res = await app.fetch(req);
+      expect(res.status).toBe(200);
 
-			const data = await res.json();
-			expect(data).toEqual({ message: "OTP sent successfully" });
-		}
+      const data = await res.json();
+      expect(data).toEqual({ message: "OTP sent successfully" });
+    }
 
-		// 4th request should be rate limited (429 status)
-		const req4 = new Request("http://localhost/v1/signers/start-onboarding", {
-			method: "POST",
-			headers: {
-				"Content-Type": "application/json",
-				authorization: env.ACCESS_SECRET,
-			},
-			body: JSON.stringify({
-				...testData,
-				deviceId: `${testData.deviceId}-4`,
-			}),
-		});
+    // 4th request should be rate limited (429 status)
+    const req4 = new Request("http://localhost/v1/signers/start-onboarding", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+        authorization: env.ACCESS_SECRET,
+      },
+      body: JSON.stringify({
+        ...testData,
+        deviceId: `${testData.deviceId}-4`,
+      }),
+    });
 
-		let res4: Response;
-		try {
-			res4 = await app.fetch(req4);
-		} catch (error) {
-			// In the test environment, the Response might be thrown
-			if (error instanceof Response) {
-				res4 = error;
-			} else {
-				throw error;
-			}
-		}
+    let res4: Response;
+    try {
+      res4 = await app.fetch(req4);
+    } catch (error) {
+      // In the test environment, the Response might be thrown
+      if (error instanceof Response) {
+        res4 = error;
+      } else {
+        throw error;
+      }
+    }
 
-		expect(res4.status).toBe(429);
+    expect(res4.status).toBe(429);
 
-		const data4 = await res4.json();
-		expect(data4.error).toContain("Too many devices onboarded recently");
-		expect(data4.retryAfterHours).toBeGreaterThan(0);
-	});
+    const data4 = await res4.json();
+    expect(data4.error).toContain("Too many devices onboarded recently");
+    expect(data4.retryAfterHours).toBeGreaterThan(0);
+  });
 });
